requirements.txt
pytest==7.1.3
pytest-xdist==2.5.0
selenium==4.5.0
webdriver-manager==3.8.3
Appium-Python-Client==2.7.1
requests==2.28.1
allure-pytest==2.11.1
SQLAlchemy==1.4.44
PyMySQL==1.0.2
cryptography==38.0.3
Faker==15.3.4
